--- src/pages/Menu/ListItems.tsx
@@ -1,26 +1,0 @@
-import React from "react";
-import { List, ListItemButton, ListItemIcon, ListItemText, Typography, Collapse, Divider } from "@mui/material";
-import { Link } from "react-router-dom";
-
-//Icons
-import ReceiptLongIcon from '@mui/icons-material/ReceiptLong';
-import LocalAtmIcon from '@mui/icons-material/LocalAtm';
-import RestorePageIcon from '@mui/icons-material/RestorePage';
-import StorefrontIcon from '@mui/icons-material/Storefront';
-import PeopleAltIcon from '@mui/icons-material/PeopleAlt';
-import AssessmentIcon from '@mui/icons-material/Assessment';
-import QuizIcon from '@mui/icons-material/Quiz';
-import ContactlessIcon from '@mui/icons-material/Contactless';
-import AccountCircleIcon from '@mui/icons-material/AccountCircle';
-import ExpandLess from '@mui/icons-material/ExpandLess';
-import ExpandMore from '@mui/icons-material/ExpandMore';
-import ArrowRightIcon from '@mui/icons-material/ArrowRight';
-
-
-
-export const ListItems = (
-    
-<React.Fragment>
-    
-</React.Fragment>
-);
